Added basic math operation support for 'add' method
JSONPointer.prototype.add now supports basic mathematical operations as an input value.  e.g: a value of '+=1' would add 1 to the existing value (or add 1 to 0 if no value exists at path).

This arguably still conforms to the RFC 6902 standards, however could be considered as a stretch on JSON API standards.